update code space
