+++ braintaichi/_misc.py
@@ -17,16 +17,16 @@ import jax
 
 
 def _get_dtype(v):
-  if hasattr(v, 'dtype'):
-    dtype = v.dtype
-  else:
-    dtype = jax.dtypes.canonicalize_dtype(type(v))
-  return dtype
+    if hasattr(v, 'dtype'):
+        dtype = v.dtype
+    else:
+        dtype = jax.dtypes.canonicalize_dtype(type(v))
+    return dtype
 
 
 def set_module_as(name: str):
-  def decorator(f):
-    f.__module__ = name
-    return f
+    def decorator(f):
+        f.__module__ = name
+        return f
 
-  return decorator
+    return decorator
