+++ braintaichi/_primitive/_batch_utils.py
@@ -23,33 +23,33 @@ from jax.interpreters import batching
 from jax.tree_util import tree_flatten, tree_unflatten
 
 __all__ = [
-  'register_general_batching',
+    'register_general_batching',
 ]
 
 
 def _general_batching_rule(prim, args, axes, **kwargs):
-  batch_axes, batch_args, non_batch_args = [], {}, {}
-  for ax_i, ax in enumerate(axes):
-    if ax is None:
-      non_batch_args[f'ax{ax_i}'] = args[ax_i]
-    else:
-      batch_args[f'ax{ax_i}'] = args[ax_i] if ax == 0 else jnp.moveaxis(args[ax_i], ax, 0)
-      batch_axes.append(ax_i)
+    batch_axes, batch_args, non_batch_args = [], {}, {}
+    for ax_i, ax in enumerate(axes):
+        if ax is None:
+            non_batch_args[f'ax{ax_i}'] = args[ax_i]
+        else:
+            batch_args[f'ax{ax_i}'] = args[ax_i] if ax == 0 else jnp.moveaxis(args[ax_i], ax, 0)
+            batch_axes.append(ax_i)
 
-  def f(_, x):
-    pars = tuple([(x[f'ax{i}'] if i in batch_axes else non_batch_args[f'ax{i}'])
-                  for i in range(len(axes))])
-    return 0, prim.bind(*pars, **kwargs)
+    def f(_, x):
+        pars = tuple([(x[f'ax{i}'] if i in batch_axes else non_batch_args[f'ax{i}'])
+                      for i in range(len(axes))])
+        return 0, prim.bind(*pars, **kwargs)
 
-  _, outs = lax.scan(f, 0, batch_args)
-  out_vals, out_tree = tree_flatten(outs)
-  out_dim = tree_unflatten(out_tree, (0,) * len(out_vals))
-  return outs, out_dim
+    _, outs = lax.scan(f, 0, batch_args)
+    out_vals, out_tree = tree_flatten(outs)
+    out_dim = tree_unflatten(out_tree, (0,) * len(out_vals))
+    return outs, out_dim
 
 
 def register_general_batching(prim):
-  batching.primitive_batchers[prim] = partial(_general_batching_rule, prim)
+    batching.primitive_batchers[prim] = partial(_general_batching_rule, prim)
 
 
 def _shape_to_layout(shape):
-  return tuple(range(len(shape) - 1, -1, -1))
+    return tuple(range(len(shape) - 1, -1, -1))
