+++ braintaichi/_sparseop/csrmv.py
@@ -27,7 +27,7 @@ from jax.interpreters import ad
 
 from braintaichi._primitive._batch_utils import register_general_batching
 from braintaichi._primitive._xla_custom_op import XLACustomOp
-from ._sparse_utils import csr_to_coo
+from .utils import csr_to_coo
 
 
 def raw_csrmv_taichi(
@@ -39,28 +39,28 @@ def raw_csrmv_taichi(
     shape: Tuple[int, int],
     transpose: bool = False,
 ):
-  out_shape = shape[1] if transpose else shape[0]
-  if data.shape[0] != 1:
-    if jax.devices()[0].platform == 'gpu':
-      return [_csr_matvec_cusparse_p.bind(data, indices, indptr, vector, shape=shape, transpose=transpose)]
+    out_shape = shape[1] if transpose else shape[0]
+    if data.shape[0] != 1:
+        if jax.devices()[0].platform == 'gpu':
+            return [_csr_matvec_cusparse_p.bind(data, indices, indptr, vector, shape=shape, transpose=transpose)]
+        else:
+            if transpose:
+                prim = _csr_matvec_transpose_heter_p
+            else:
+                prim = _csr_matvec_heter_p
     else:
-      if transpose:
-        prim = _csr_matvec_transpose_heter_p
-      else:
-        prim = _csr_matvec_heter_p
-  else:
-    if transpose:
-      prim = _csr_matvec_transpose_homo_p
-    else:
-      prim = _csr_matvec_homo_p
+        if transpose:
+            prim = _csr_matvec_transpose_homo_p
+        else:
+            prim = _csr_matvec_homo_p
 
-  return prim(data,
-              indices,
-              indptr,
-              vector,
-              outs=[jax.ShapeDtypeStruct((out_shape,), dtype=data.dtype)],
-              transpose=transpose,
-              shape=shape)
+    return prim(data,
+                indices,
+                indptr,
+                vector,
+                outs=[jax.ShapeDtypeStruct((out_shape,), dtype=data.dtype)],
+                transpose=transpose,
+                shape=shape)
 
 
 # -------------
@@ -72,11 +72,11 @@ def _sparse_csr_matvec_transpose_homo_cpu(values: ti.types.ndarray(ndim=1),
                                           row_ptr: ti.types.ndarray(ndim=1),
                                           vector: ti.types.ndarray(ndim=1),
                                           out: ti.types.ndarray(ndim=1)):
-  value = values[0]
-  ti.loop_config(serialize=True)
-  for row_i in range(row_ptr.shape[0] - 1):
-    for j in range(row_ptr[row_i], row_ptr[row_i + 1]):
-      out[col_indices[j]] += value * vector[row_i]
+    value = values[0]
+    ti.loop_config(serialize=True)
+    for row_i in range(row_ptr.shape[0] - 1):
+        for j in range(row_ptr[row_i], row_ptr[row_i + 1]):
+            out[col_indices[j]] += value * vector[row_i]
 
 
 @ti.kernel
@@ -85,10 +85,10 @@ def _sparse_csr_matvec_transpose_heter_cpu(values: ti.types.ndarray(ndim=1),
                                            row_ptr: ti.types.ndarray(ndim=1),
                                            vector: ti.types.ndarray(ndim=1),
                                            out: ti.types.ndarray(ndim=1)):
-  ti.loop_config(serialize=True)
-  for row_i in range(row_ptr.shape[0] - 1):
-    for j in range(row_ptr[row_i], row_ptr[row_i + 1]):
-      out[col_indices[j]] += vector[row_i] * values[j]
+    ti.loop_config(serialize=True)
+    for row_i in range(row_ptr.shape[0] - 1):
+        for j in range(row_ptr[row_i], row_ptr[row_i + 1]):
+            out[col_indices[j]] += vector[row_i] * values[j]
 
 
 @ti.kernel
@@ -97,13 +97,13 @@ def _sparse_csr_matvec_homo_cpu(values: ti.types.ndarray(ndim=1),
                                 row_ptr: ti.types.ndarray(ndim=1),
                                 vector: ti.types.ndarray(ndim=1),
                                 out: ti.types.ndarray(ndim=1)):
-  value = values[0]
-  # ti.loop_config(serialize=True)
-  for row_i in range(row_ptr.shape[0] - 1):
-    r = 0.
-    for j in range(row_ptr[row_i], row_ptr[row_i + 1]):
-      r += vector[col_indices[j]]
-    out[row_i] = r * value
+    value = values[0]
+    # ti.loop_config(serialize=True)
+    for row_i in range(row_ptr.shape[0] - 1):
+        r = 0.
+        for j in range(row_ptr[row_i], row_ptr[row_i + 1]):
+            r += vector[col_indices[j]]
+        out[row_i] = r * value
 
 
 @ti.kernel
@@ -112,12 +112,12 @@ def _sparse_csr_matvec_heter_cpu(values: ti.types.ndarray(ndim=1),
                                  row_ptr: ti.types.ndarray(ndim=1),
                                  vector: ti.types.ndarray(ndim=1),
                                  out: ti.types.ndarray(ndim=1)):
-  # ti.loop_config(serialize=True)
-  for row_i in range(row_ptr.shape[0] - 1):
-    r = 0.
-    for j in range(row_ptr[row_i], row_ptr[row_i + 1]):
-      r += values[j] * vector[col_indices[j]]
-    out[row_i] = r
+    # ti.loop_config(serialize=True)
+    for row_i in range(row_ptr.shape[0] - 1):
+        r = 0.
+        for j in range(row_ptr[row_i], row_ptr[row_i + 1]):
+            r += values[j] * vector[col_indices[j]]
+        out[row_i] = r
 
 
 # -------------
@@ -130,15 +130,15 @@ def _sparse_csr_matvec_transpose_homo_gpu(values: ti.types.ndarray(ndim=1),
                                           row_ptr: ti.types.ndarray(ndim=1),
                                           vector: ti.types.ndarray(ndim=1),
                                           out: ti.types.ndarray(ndim=1)):
-  value = values[0]
-  for i in range((row_ptr.shape[0] - 1) * 32):
-    row_i = i >> 5
-    index = i & 31
-    j = row_ptr[row_i] + index
-    end_index = row_ptr[row_i + 1]
-    while j < end_index:
-      out[col_indices[j]] += value * vector[row_i]
-      j += 32
+    value = values[0]
+    for i in range((row_ptr.shape[0] - 1) * 32):
+        row_i = i >> 5
+        index = i & 31
+        j = row_ptr[row_i] + index
+        end_index = row_ptr[row_i + 1]
+        while j < end_index:
+            out[col_indices[j]] += value * vector[row_i]
+            j += 32
 
 
 @ti.kernel
@@ -147,17 +147,17 @@ def _sparse_csr_matvec_homo_gpu(values: ti.types.ndarray(ndim=1),
                                 row_ptr: ti.types.ndarray(ndim=1),
                                 vector: ti.types.ndarray(ndim=1),
                                 out: ti.types.ndarray(ndim=1)):
-  value = values[0]
-  for i in range((row_ptr.shape[0] - 1) * 32):
-    row_i = i >> 5
-    index = i & 31
-    r = 0.
-    j = row_ptr[row_i] + index
-    end_index = row_ptr[row_i + 1]
-    while j < end_index:
-      r += vector[col_indices[j]]
-      j += 32
-    out[row_i] += value * r
+    value = values[0]
+    for i in range((row_ptr.shape[0] - 1) * 32):
+        row_i = i >> 5
+        index = i & 31
+        r = 0.
+        j = row_ptr[row_i] + index
+        end_index = row_ptr[row_i + 1]
+        while j < end_index:
+            r += vector[col_indices[j]]
+            j += 32
+        out[row_i] += value * r
 
 
 @ti.kernel
@@ -166,14 +166,14 @@ def _sparse_csr_matvec_transpose_heter_gpu(values: ti.types.ndarray(ndim=1),
                                            row_ptr: ti.types.ndarray(ndim=1),
                                            vector: ti.types.ndarray(ndim=1),
                                            out: ti.types.ndarray(ndim=1)):
-  for i in range((row_ptr.shape[0] - 1) * 32):
-    row_i = i >> 5
-    index = i & 31
-    j = row_ptr[row_i] + index
-    end_index = row_ptr[row_i + 1]
-    while j < end_index:
-      out[col_indices[j]] += values[j] * vector[row_i]
-      j += 32
+    for i in range((row_ptr.shape[0] - 1) * 32):
+        row_i = i >> 5
+        index = i & 31
+        j = row_ptr[row_i] + index
+        end_index = row_ptr[row_i + 1]
+        while j < end_index:
+            out[col_indices[j]] += values[j] * vector[row_i]
+            j += 32
 
 
 @ti.kernel
@@ -182,54 +182,54 @@ def _sparse_csr_matvec_heter_gpu(values: ti.types.ndarray(ndim=1),
                                  row_ptr: ti.types.ndarray(ndim=1),
                                  vector: ti.types.ndarray(ndim=1),
                                  out: ti.types.ndarray(ndim=1)):
-  for i in range((row_ptr.shape[0] - 1) * 32):
-    row_i = i >> 5
-    index = i & 31
-    r = 0.
-    j = row_ptr[row_i] + index
-    end_index = row_ptr[row_i + 1]
-    while j < end_index:
-      r += values[j] * vector[col_indices[j]]
-      j += 32
-    out[row_i] += r  # TODO: warp-level primitive
+    for i in range((row_ptr.shape[0] - 1) * 32):
+        row_i = i >> 5
+        index = i & 31
+        r = 0.
+        j = row_ptr[row_i] + index
+        end_index = row_ptr[row_i + 1]
+        while j < end_index:
+            r += values[j] * vector[col_indices[j]]
+            j += 32
+        out[row_i] += r  # TODO: warp-level primitive
 
 
 def _sparse_csr_matvec_jvp_values(val_dot, values, col_indices, row_ptr, vector, *, outs, transpose, shape):
-  return raw_csrmv_taichi(val_dot, col_indices, row_ptr, vector, shape=shape, transpose=transpose)
+    return raw_csrmv_taichi(val_dot, col_indices, row_ptr, vector, shape=shape, transpose=transpose)
 
 
 def _sparse_csr_matvec_jvp_vector(vec_dot, values, col_indices, row_ptr, vector, *, outs, transpose, shape):
-  return raw_csrmv_taichi(values, col_indices, row_ptr, vec_dot, shape=shape, transpose=transpose)
+    return raw_csrmv_taichi(values, col_indices, row_ptr, vec_dot, shape=shape, transpose=transpose)
 
 
 def _sparse_csr_matvec_transpose(
     ct, data, indices, indptr, vector, *, outs, transpose, shape,
 ):
-  if ad.is_undefined_primal(indices) or ad.is_undefined_primal(indptr):
-    raise ValueError("Cannot transpose with respect to sparse indices.")
-  if ad.is_undefined_primal(vector):
-    ct_vector = raw_csrmv_taichi(data, indices, indptr, ct[0], shape=shape, transpose=not transpose)[0]
-    return data, indices, indptr, (ad.Zero(vector) if type(ct[0]) is ad.Zero else ct_vector)
+    if ad.is_undefined_primal(indices) or ad.is_undefined_primal(indptr):
+        raise ValueError("Cannot transpose with respect to sparse indices.")
+    if ad.is_undefined_primal(vector):
+        ct_vector = raw_csrmv_taichi(data, indices, indptr, ct[0], shape=shape, transpose=not transpose)[0]
+        return data, indices, indptr, (ad.Zero(vector) if type(ct[0]) is ad.Zero else ct_vector)
 
-  else:
-    if type(ct[0]) is ad.Zero:
-      ct_data = ad.Zero(data)
     else:
-      if data.aval.shape[0] == 1:  # scalar
-        ct_data = raw_csrmv_taichi(jnp.ones(1), indices, indptr, vector, shape=shape, transpose=transpose)[0]
-        ct_data = jnp.inner(ct[0], ct_data)
-      else:
-        row, col = csr_to_coo(indices, indptr)
-        ct_data = vector[row] * ct[0][col] if transpose else vector[col] * ct[0][row]
+        if type(ct[0]) is ad.Zero:
+            ct_data = ad.Zero(data)
+        else:
+            if data.aval.shape[0] == 1:  # scalar
+                ct_data = raw_csrmv_taichi(jnp.ones(1), indices, indptr, vector, shape=shape, transpose=transpose)[0]
+                ct_data = jnp.inner(ct[0], ct_data)
+            else:
+                row, col = csr_to_coo(indices, indptr)
+                ct_data = vector[row] * ct[0][col] if transpose else vector[col] * ct[0][row]
 
-    return ct_data, indices, indptr, vector
+        return ct_data, indices, indptr, vector
 
 
 def _define_op(cpu_kernel, gpu_kernel):
-  prim = XLACustomOp(cpu_kernel=cpu_kernel, gpu_kernel=gpu_kernel)
-  prim.defjvp(_sparse_csr_matvec_jvp_values, None, None, _sparse_csr_matvec_jvp_vector)
-  prim.def_transpose_rule(_sparse_csr_matvec_transpose)
-  return prim
+    prim = XLACustomOp(cpu_kernel=cpu_kernel, gpu_kernel=gpu_kernel)
+    prim.defjvp(_sparse_csr_matvec_jvp_values, None, None, _sparse_csr_matvec_jvp_vector)
+    prim.def_transpose_rule(_sparse_csr_matvec_transpose)
+    return prim
 
 
 # transpose homo
