+++ braintaichi/_sparseop/main.py
@@ -20,14 +20,14 @@ import jax
 from jax import numpy as jnp, dtypes, default_backend
 
 from braintaichi._misc import set_module_as
-from ._sparse_coomv import _coomv_cusparse_p
-from ._sparse_csrmm import raw_csrmm_taichi
-from ._sparse_csrmv import raw_csrmv_taichi
+from .coomv import _coomv_cusparse_p
+from .csrmm import raw_csrmm_taichi
+from .csrmv import raw_csrmv_taichi
 
 __all__ = [
-  'coomv',
-  'csrmv',
-  'csrmm',
+    'coomv',
+    'csrmv',
+    'csrmm',
 ]
 
 
@@ -44,65 +44,65 @@ def coomv(
     transpose: bool = False,
     method: str = 'cusparse'
 ):
-  """Product of COO sparse matrix and a dense vector using cuSPARSE algorithm.
+    """Product of COO sparse matrix and a dense vector using cuSPARSE algorithm.
 
-  This function supports JAX transformations, including `jit()`, `grad()`,
-  `vmap()` and `pmap()`.
+    This function supports JAX transformations, including `jit()`, `grad()`,
+    `vmap()` and `pmap()`.
 
-  Parameters
-  ----------
-  data: ndarray, float
-    An array of shape ``(nse,)``.
-  row: ndarray
-    An array of shape ``(nse,)``.
-  col: ndarray
-    An array of shape ``(nse,)`` and dtype ``row.dtype``.
-  vector: ndarray
-    An array of shape ``(shape[0] if transpose else shape[1],)`` and
-    dtype ``data.dtype``.
-  shape: tuple of int
-    The shape of the sparse matrix.
-  rows_sorted: bool
-    Row index are sorted.
-  cols_sorted: bool
-    Column index are sorted.
-  transpose: bool
-    A boolean specifying whether to transpose the sparse matrix
-    before computing.
-  method: str
-    The method used to compute the matrix-vector multiplication.
+    Parameters
+    ----------
+    data: ndarray, float
+      An array of shape ``(nse,)``.
+    row: ndarray
+      An array of shape ``(nse,)``.
+    col: ndarray
+      An array of shape ``(nse,)`` and dtype ``row.dtype``.
+    vector: ndarray
+      An array of shape ``(shape[0] if transpose else shape[1],)`` and
+      dtype ``data.dtype``.
+    shape: tuple of int
+      The shape of the sparse matrix.
+    rows_sorted: bool
+      Row index are sorted.
+    cols_sorted: bool
+      Column index are sorted.
+    transpose: bool
+      A boolean specifying whether to transpose the sparse matrix
+      before computing.
+    method: str
+      The method used to compute the matrix-vector multiplication.
 
-  Returns
-  -------
-  y: ndarray
-    An array of shape ``(shape[1] if transpose else shape[0],)`` representing
-    the matrix vector product.
-  """
+    Returns
+    -------
+    y: ndarray
+      An array of shape ``(shape[1] if transpose else shape[0],)`` representing
+      the matrix vector product.
+    """
 
-  data = jnp.atleast_1d(jnp.asarray(data))
-  row = jnp.asarray(row)
-  col = jnp.asarray(col)
-  vector = jnp.asarray(vector)
+    data = jnp.atleast_1d(jnp.asarray(data))
+    row = jnp.asarray(row)
+    col = jnp.asarray(col)
+    vector = jnp.asarray(vector)
 
-  if method == 'cusparse':
-    if default_backend() != 'cpu':
-      if data.shape[0] == 1:
-        data = jnp.ones(row.shape, dtype=data.dtype) * data
-      if row.dtype in [jnp.uint32, jnp.uint64]:
-        row = jnp.asarray(row, dtype=dtypes.canonicalize_dtype(jnp.int64))
-      if col.dtype in [jnp.uint32, jnp.uint64]:
-        col = jnp.asarray(col, dtype=dtypes.canonicalize_dtype(jnp.int64))
-    return _coomv_cusparse_p.bind(data,
-                                  row,
-                                  col,
-                                  vector,
-                                  shape=shape,
-                                  rows_sorted=rows_sorted,
-                                  cols_sorted=cols_sorted,
-                                  transpose=transpose)
+    if method == 'cusparse':
+        if default_backend() != 'cpu':
+            if data.shape[0] == 1:
+                data = jnp.ones(row.shape, dtype=data.dtype) * data
+            if row.dtype in [jnp.uint32, jnp.uint64]:
+                row = jnp.asarray(row, dtype=dtypes.canonicalize_dtype(jnp.int64))
+            if col.dtype in [jnp.uint32, jnp.uint64]:
+                col = jnp.asarray(col, dtype=dtypes.canonicalize_dtype(jnp.int64))
+        return _coomv_cusparse_p.bind(data,
+                                      row,
+                                      col,
+                                      vector,
+                                      shape=shape,
+                                      rows_sorted=rows_sorted,
+                                      cols_sorted=cols_sorted,
+                                      transpose=transpose)
 
-  else:
-    raise ValueError
+    else:
+        raise ValueError
 
 
 @set_module_as('braintaichi')
@@ -115,24 +115,24 @@ def csrmm(
     shape: Tuple[int, int],
     transpose: bool = False,
 ):
-  """
-  Product of CSR sparse matrix and a dense matrix.
+    """
+    Product of CSR sparse matrix and a dense matrix.
 
-  Args:
-      data : array of shape ``(nse,)``.
-      indices : array of shape ``(nse,)``
-      indptr : array of shape ``(shape[0] + 1,)`` and dtype ``indices.dtype``
-      B : array of shape ``(shape[0] if transpose else shape[1], cols)`` and
-      dtype ``data.dtype``
-      shape : length-2 tuple representing the matrix shape
-      transpose : boolean specifying whether to transpose the sparse matrix
-      before computing.
+    Args:
+        data : array of shape ``(nse,)``.
+        indices : array of shape ``(nse,)``
+        indptr : array of shape ``(shape[0] + 1,)`` and dtype ``indices.dtype``
+        B : array of shape ``(shape[0] if transpose else shape[1], cols)`` and
+        dtype ``data.dtype``
+        shape : length-2 tuple representing the matrix shape
+        transpose : boolean specifying whether to transpose the sparse matrix
+        before computing.
 
-  Returns:
-      C : array of shape ``(shape[1] if transpose else shape[0], cols)``
-      representing the matrix-matrix product.
-  """
-  return raw_csrmm_taichi(data, indices, indptr, matrix, shape=shape, transpose=transpose)[0]
+    Returns:
+        C : array of shape ``(shape[1] if transpose else shape[0], cols)``
+        representing the matrix-matrix product.
+    """
+    return raw_csrmm_taichi(data, indices, indptr, matrix, shape=shape, transpose=transpose)[0]
 
 
 @set_module_as('braintaichi')
@@ -145,63 +145,63 @@ def csrmv(
     shape: Tuple[int, int],
     transpose: bool = False,
 ):
-  """Product of CSR sparse matrix and a dense vector using cuSPARSE algorithm.
+    """Product of CSR sparse matrix and a dense vector using cuSPARSE algorithm.
 
-  This function supports JAX transformations, including `jit()`, `grad()`,
-  `vmap()` and `pmap()`.
+    This function supports JAX transformations, including `jit()`, `grad()`,
+    `vmap()` and `pmap()`.
 
-  Parameters
-  ----------
-  data: ndarray, float
-    An array of shape ``(nse,)``.
-  indices: ndarray
-    An array of shape ``(nse,)``.
-  indptr: ndarray
-    An array of shape ``(shape[0] + 1,)`` and dtype ``indices.dtype``.
-  vector: ndarray
-    An array of shape ``(shape[0] if transpose else shape[1],)``
-    and dtype ``data.dtype``.
-  shape: tuple of int
-    A length-2 tuple representing the matrix shape.
-  transpose: bool
-    A boolean specifying whether to transpose the sparse matrix
-    before computing.
-  method: str
-    The method used to compute Matrix-Vector Multiplication. Default is ``taichi``.
-    The candidate methods are:
+    Parameters
+    ----------
+    data: ndarray, float
+      An array of shape ``(nse,)``.
+    indices: ndarray
+      An array of shape ``(nse,)``.
+    indptr: ndarray
+      An array of shape ``(shape[0] + 1,)`` and dtype ``indices.dtype``.
+    vector: ndarray
+      An array of shape ``(shape[0] if transpose else shape[1],)``
+      and dtype ``data.dtype``.
+    shape: tuple of int
+      A length-2 tuple representing the matrix shape.
+    transpose: bool
+      A boolean specifying whether to transpose the sparse matrix
+      before computing.
+    method: str
+      The method used to compute Matrix-Vector Multiplication. Default is ``taichi``.
+      The candidate methods are:
 
-    - ``None``: default using Taichi kernel.
-    - ``cusparse``: using cuSPARSE library.
-    - ``scalar``:
-    - ``vector``:
-    - ``adaptive``:
+      - ``None``: default using Taichi kernel.
+      - ``cusparse``: using cuSPARSE library.
+      - ``scalar``:
+      - ``vector``:
+      - ``adaptive``:
 
-  Returns
-  -------
-  y : ndarry
-    The array of shape ``(shape[1] if transpose else shape[0],)`` representing
-    the matrix vector product.
-  """
+    Returns
+    -------
+    y : ndarry
+      The array of shape ``(shape[1] if transpose else shape[0],)`` representing
+      the matrix vector product.
+    """
 
-  data = jnp.atleast_1d(data)
+    data = jnp.atleast_1d(data)
 
-  if vector.dtype == jnp.bool_:
-    vector = jnp.asarray(vector, dtype=data.dtype)
+    if vector.dtype == jnp.bool_:
+        vector = jnp.asarray(vector, dtype=data.dtype)
 
-  if data.dtype not in [jnp.float16, jnp.float32, jnp.float64]:
-    raise TypeError('Only support float16, float32 or float64 type. '
-                    f'But we got {data.dtype}.')
-  if data.dtype != vector.dtype:
-    raise TypeError('The types of data and vector should be the same. '
-                    f'But we got {data.dtype} != {vector.dtype}.')
-  assert data.ndim == indices.ndim == indptr.ndim == vector.ndim == 1
-  if not jnp.issubdtype(indices.dtype, jnp.integer):
-    raise ValueError('indices should be a 1D vector with integer type.')
-  if not jnp.issubdtype(indptr.dtype, jnp.integer):
-    raise ValueError('indptr should be a 1D vector with integer type.')
+    if data.dtype not in [jnp.float16, jnp.float32, jnp.float64]:
+        raise TypeError('Only support float16, float32 or float64 type. '
+                        f'But we got {data.dtype}.')
+    if data.dtype != vector.dtype:
+        raise TypeError('The types of data and vector should be the same. '
+                        f'But we got {data.dtype} != {vector.dtype}.')
+    assert data.ndim == indices.ndim == indptr.ndim == vector.ndim == 1
+    if not jnp.issubdtype(indices.dtype, jnp.integer):
+        raise ValueError('indices should be a 1D vector with integer type.')
+    if not jnp.issubdtype(indptr.dtype, jnp.integer):
+        raise ValueError('indptr should be a 1D vector with integer type.')
 
-  # if the shape of indices is (0,), then we return a zero vector
-  if indices.shape[0] == 0:
-    return jnp.zeros(shape[1] if transpose else shape[0], dtype=data.dtype)
+    # if the shape of indices is (0,), then we return a zero vector
+    if indices.shape[0] == 0:
+        return jnp.zeros(shape[1] if transpose else shape[0], dtype=data.dtype)
 
-  return raw_csrmv_taichi(data, indices, indptr, vector, shape=shape, transpose=transpose)[0]
+    return raw_csrmv_taichi(data, indices, indptr, vector, shape=shape, transpose=transpose)[0]
