+++ braintaichi/rand/__init__.py
@@ -12,12 +12,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-
-from .utils import *
-from .utils import __all__ as _sparse_utils_all
-from .main import *
-from .main import __all__ as _main_all
-
-__all__ = _main_all + _sparse_utils_all
-
-del _sparse_utils_all, _main_all
